results of a major netbeans inspect & transform

+++ src/main/java/org/encog/ca/universe/basic/BasicDiscreteCell.java
@@ -74,7 +74,7 @@ public class BasicDiscreteCell implements DiscreteCell, Serializable {
 
 	@Override
 	public double getAvg() {
-		return (int) EngineArray.mean(this.data);
+		return EngineArray.mean(this.data);
 	}
 
 	@Override
